# ai-services/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for managing AI Services catalog, applications, and authentication",
        "title": "AI Services Catalog API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/project-ai-services/ai-services",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/architectures": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all available architecture templates with summary information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "List available architectures",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ArchitectureSummary"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/architectures/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves detailed information about a specific architecture template",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Get architecture details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Architecture template ID (e.g., 'rag')",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.Architecture"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Architecture not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/architectures/{id}/deploy-options": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves available providers and dependency rules for all services and their components within an architecture",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Get architecture deploy options",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Architecture ID (e.g., 'rag')",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.DeployOptionsArchitecture"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Architecture not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/components/{component_type}/providers/{provider_id}/params": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the configuration schema (JSON Schema) for a specific provider within a component type. Returns a JSON Schema object with properties that may include x-data-id for fields that should be populated from metadata specifications.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Get component provider parameters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Component type (e.g., 'vector_db', 'llm', 'embedding', 'reranker')",
                        "name": "component_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Provider identifier (e.g., 'opensearch', 'vllm', 'watsonx')",
                        "name": "provider_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON Schema object with $schema, type, and properties. Properties may include x-data-id field indicating data should be populated from metadata specifications (e.g., supported_models)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid component_type or provider_id",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Component type or provider not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/services": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all deployable service templates. Dependency-only services are excluded from this list. Returns service summaries including standalone flag without endpoints and pod templates.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "List available services",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ServiceSummary"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/services/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves detailed information about a specific service template",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Get service details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service template ID (e.g., 'summarize')",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.Service"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Service not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/services/{id}/deploy-options": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves available providers and dependency rules for a specific service",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Get service deploy options",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service ID (e.g., 'digitize', 'chat')",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.DeployOptionsService"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Service not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/services/{id}/params": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the configuration schema (JSON Schema) for a specific service. Returns a JSON Schema object with properties that define the service's configurable parameters.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Catalog"
                ],
                "summary": "Get service parameters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Service ID (e.g., 'chat', 'digitize', 'similarity')",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON Schema object with $schema, type, and properties defining service parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid service ID",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing access token",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Service not found",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_pkg_catalog_apiserver_handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.Architecture": {
            "type": "object",
            "properties": {
                "certified_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "global_components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ComponentReference"
                    }
                },
                "id": {
                    "type": "string"
                },
                "links": {
                    "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ArchitectureLinks"
                },
                "name": {
                    "type": "string"
                },
                "runtimes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ServiceReference"
                    }
                },
                "type": {
                    "description": "\"architecture\"",
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ArchitectureLinks": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "demo": {
                    "type": "string"
                },
                "documentation": {
                    "type": "string"
                }
            }
        },
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ArchitectureSummary": {
            "type": "object",
            "properties": {
                "certified_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ComponentReference": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                }
            }
        },
...
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.DeployOptionsArchitecture": {
            "type": "object",
            "properties": {
                "global_components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.DeployOptionsComponent"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.DeployOptionsService"
                    }
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.DeployOptionsService": {
            "type": "object",
            "properties": {
                "components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.DeployOptionsComponent"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "resources": {
                    "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.Resources"
                },
                "schema": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.Service": {
            "type": "object",
            "properties": {
                "architectures": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "certified_by": {
                    "type": "string"
                },
                "dependencies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_project-ai-services_ai-services_internal_pkg_catalog_types.DependencyReference"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "standalone": {
                    "type": "boolean"
                },
                "type": {
                    "description": "\"service\"",
                    "type": "string"
                }
            }
        },
...
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ServiceReference": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "optional": {
                    "type": "boolean"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "github_com_project-ai-services_ai-services_internal_pkg_catalog_types.ServiceSummary": {
            "type": "object",
            "properties": {
                "architectures": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "certified_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "standalone": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_pkg_catalog_apiserver_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}